order_service/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
pika==1.3.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
drf_yasg==1.21.7
django_celery_results==2.5.1
